Revert "Change to lowercase"
