+++ .eslintignore
@@ -2,4 +2,3 @@
 .eslintrc.js
 jest.config.js
 jest.setup.js
-webpack.config.js
