+++ src/pages/Content/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
-import App from './app';
+import App from './App';
 
 console.log('Content script works!');
