+++ src/pages/Newtab/Newtab.css
