+++ src/pages/Newtab/Newtab.scss
