+++ src/pages/Options/Options.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './options.css';
+import './Options.css';
 
 interface Props {
   title: string;
