+++ src/pages/Options/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Options from './options';
+import Options from './Options';
 import './index.css';
 
 render(
