+++ src/pages/Panel/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Panel from './panel';
+import Panel from './Panel';
 import './index.css';
 
 render(<Panel />, window.document.querySelector('#app-container'));
