+++ src/pages/Popup/Popup.jsx
@@ -1,19 +1,14 @@
 import React from 'react';
 import logo from '../../assets/runtime/logo.svg';
-import './newtab.css';
-import './newtab.scss';
+import './Popup.css';
 
-export default function Newtab() {
+function Popup() {
   return (
     <div className="App">
       <header className="App-header">
-        <img
-          src={chrome.runtime.getURL('logo.svg')}
-          className="App-logo"
-          alt="logo"
-        />
+        <img src={logo} className="App-logo" alt="logo" />
         <p>
-          Edit <code>src/pages/Newtab/Newtab.js</code> and save to reload.
+          Edit <code>src/pages/Popup/Popup.jsx</code> and save to reload.
         </p>
         <a
           className="App-link"
@@ -23,8 +18,9 @@ export default function Newtab() {
         >
           Learn React!
         </a>
-        <h6>The color of this paragraph is defined using SASS.</h6>
       </header>
     </div>
   );
 }
+
+export default Popup;
