+++ src/pages/Popup/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Popup from './popup';
+import Popup from './Popup';
 import './index.css';
 
 render(<Popup />, window.document.querySelector('#app-container'));
